rts-tf2/requirements.txt
tensorflow==2.0.0rc1
tensorflow-addons
git+https://github.com/tensorflow/examples.git#egg=tensorflow_examples
plac
numpy
matplotlib
jupyter
pillow
gast==0.2.2